window-slot: Jump to parent also for non-native locations
When the current location is marked as gone, Nautilus jumps to the
first existing parent currently (except when unmounted), but only for
native locations. I don't see reasons why not to jump to parent also
for GVfs locations when they are still mounted (though not many of them
support file monitors). Let's change that.