window-slot: Jump to home when current location was unmounted
When the currently displayed location is removed, we automatically jump
to the parent directory, or, if that's gone too, its parent, and so on.

Except for `GMount` root folders: it's not desirable to go up to the
parent of the mountpoint, so we jump to Home instead. For this we rely
on the `GMount` still being around.

However, for subfolders it still jumps to the first existing parent.
And the same trick doesn't work so well here: the GMount may already
be gone.

Let's instead check whether the location has been unmounted.