directory-async: Don't skip files in deep count/size
We always skip hidden and backup files when counting the number of
files in a directory, if "show-hidden-files" is turned Off. This was
done so that the item count matches which items the user sees.

However, we have been skipping them also when computing the total size
of a folder contents. This is wrong because hiding content doesn't
make the folder actually smaller in size.

So, let's always count hidden and backup files for total folder size,
and, for consistency, also for deep count.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/638